% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptspec.R
\name{adaptspec_tuning}
\alias{adaptspec_tuning}
\title{Tuning the AdaptSPEC sampler}
\usage{
adaptspec_tuning(
  prob_short_move = 0.8,
  short_moves = c(-1, 0, 1),
  short_move_weights = c(0.5, 0.5, 0.5),
  var_inflate = 1,
  warm_up_var_inflate = var_inflate,
  use_cutpoint_within = TRUE,
  use_single_within = TRUE,
  use_hmc_within = FALSE,
  l_min = 1,
  l_max = 10,
  epsilon_min = 0.1,
  epsilon_max = 1,
  use_hessian_curvature = TRUE
)
}
\arguments{
\item{prob_short_move}{probability of proposing a small move of a cutpoint
in a within step}

\item{short_moves}{set of possible small steps for the 'small move'
proposal. if using time_step > 1, will be multiplied by time_step.}

\item{short_move_weights}{the probability of picking each corresponding
entry in \code{short_moves}. Need not be normalised to sum to one}

\item{var_inflate}{factor by which to inflate the covariance matrix of the
spline coefficient proposal}

\item{warm_up_var_inflate}{as above, but applying only during warm-up}

\item{use_cutpoint_within}{whether to propose the cutpoints. One would very
rarely want to disable this.}

\item{use_single_within}{whether to also update spline directly, not just
when updating cutpoints}

\item{use_hmc_within}{whether to also update spline parameters using HMC}

\item{l_min}{minimum number of leap frog steps to take in an HMC iteration.
Number of steps is randomly selected from a discrete uniform over
\code{[l_min, l_max]}}

\item{l_max}{maximum number of lead frog steps to take in an HMC iteration}

\item{epsilon_min}{minimum step size in HMC. Step size is chosen from a
uniform distribution over \code{[epsilon_min, epsilon_max]}}

\item{epsilon_max}{maximum step size in HMC.}

\item{use_hessian_curvature}{whether to use the hessian or the Fisher
information for curvature of spline fit}
}
\description{
This specifies the tuning parameters of the reversible-jump MCMC algorithm
used to estimate AdaptSPEC. The output is a list.
}
