% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptspecfit.R
\name{adaptspecfit}
\alias{adaptspecfit}
\alias{window.adaptspecfit}
\alias{summary.adaptspecfit}
\alias{ntimes.adaptspecfit}
\alias{diagnostic_plots.adaptspecfit}
\alias{diagnostics.adaptspecfit}
\alias{diagnostic_warnings.adaptspecfit}
\alias{plot.adaptspecfit}
\title{Methods for adaptspecfit objects}
\usage{
\method{window}{adaptspecfit}(x, ...)

\method{summary}{adaptspecfit}(object, iterations_threshold = 0, ...)

\method{ntimes}{adaptspecfit}(x, ...)

\method{diagnostic_plots}{adaptspecfit}(x, ...)

\method{diagnostics}{adaptspecfit}(x, iterations_threshold = 0, ...)

\method{diagnostic_warnings}{adaptspecfit}(
  x,
  effective_size_threshold = 100,
  iterations_proportion_threshold = 0.01,
  prefix = "",
  ...
)

\method{plot}{adaptspecfit}(x, ask, auto_layout = TRUE, ...)
}
\arguments{
\item{x, object}{\code{adaptspecfit} object}

\item{...}{Extra arguments. For \code{window}, these are passed to
\code{\link[coda]{window.mcmc}} to set the \code{thin} or \code{start}
parameters (for example)}

\item{iterations_threshold}{Number of iterations below which to suppress
output. For example, if this is 100, and the MCMC chain spent less than
100 iterations in, say, n_segments = 4, summary output will not include
n_segments = 4}

\item{effective_size_threshold}{Minimum effective sample size (see
\code{\link[coda]{effectiveSize}}) below which a warning is output.}

\item{iterations_proportion_threshold}{Proportion of iterations below which
to suppress output. Just like \code{iterations_threshold}, but a proportion.}

\item{prefix}{Used internally.}

\item{ask}{Prompt user before each page of plots.}

\item{auto_layout}{Automatically generate output format.}
}
\description{
These methods apply to the adaptspecfit objects returned by
\code{\link{adaptspec}}.
}
\section{Functions}{
\itemize{
\item \code{window.adaptspecfit}: Method to modify the start/thinning of MCMC samples,
as per \code{\link[coda]{window.mcmc}}

\item \code{summary.adaptspecfit}: Summarises MCMC samples.

\item \code{ntimes.adaptspecfit}: Returns the number of time periods in the input

\item \code{diagnostic_plots.adaptspecfit}: MCMC diagnostic plots to assess convergence.

\item \code{diagnostics.adaptspecfit}: Outputs MCMC diagnostics statistics to help assess
convergence.

\item \code{diagnostic_warnings.adaptspecfit}: Outputs warnings when MCMC diagnostics are below
outside of nominal threshold ranges. The user is cautioned that this is
subject to both false positive and false negatives; examining diagnostics
plots directly is advised.

\item \code{plot.adaptspecfit}: Summary plots
}}

\seealso{
\code{\link{segment_log_spectra_mean}} for methods calculating
estimates of the spectral densities in each segments.
\code{\link{time_varying_mean_mean.adaptspecfit}} for estimates of the
time-varying mean and spectral density.
}
