% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptspecmixturefit.R
\name{adaptspecmixturefit}
\alias{adaptspecmixturefit}
\alias{window.adaptspecmixturefit}
\alias{ntimes.adaptspecmixturefit}
\alias{diagnostics.adaptspecmixturefit}
\alias{diagnostic_plots.adaptspecmixturefit}
\alias{diagnostic_warnings.adaptspecmixturefit}
\title{Methods for adaptspecmixturefit objects}
\usage{
\method{window}{adaptspecmixturefit}(x, ...)

\method{ntimes}{adaptspecmixturefit}(x, ...)

\method{diagnostics}{adaptspecmixturefit}(x, ...)

\method{diagnostic_plots}{adaptspecmixturefit}(x, top = NULL, ...)

\method{diagnostic_warnings}{adaptspecmixturefit}(x, ...)
}
\arguments{
\item{x}{\code{adaptspecmixturefit} object}

\item{...}{Extra arguments. For \code{window}, these are passed to
\code{\link[coda]{window.mcmc}} to set the \code{thin} or \code{start}
parameters (for example)}

\item{top}{Text to show in the heading of the plot}
}
\description{
These methods apply to the adaptspecmixturefit objects returned
by \code{\link{adaptspecx}}.
}
\section{Functions}{
\itemize{
\item \code{window.adaptspecmixturefit}: Method to modify the start/thinning of MCMC
samples, as per \code{\link[coda]{window.mcmc}}

\item \code{ntimes.adaptspecmixturefit}: Returns the number of time periods in the
input

\item \code{diagnostics.adaptspecmixturefit}: Outputs MCMC diagnostics statistics to help
assess convergence. Calls \code{\link{diagnostics.adaptspecfit}}.

\item \code{diagnostic_plots.adaptspecmixturefit}: MCMC diagnostic plots to assess convergence.
Calls \code{\link{diagnostics.adaptspecfit}}.

\item \code{diagnostic_warnings.adaptspecmixturefit}: Outputs warnings when MCMC diagnostics are
below outside of nominal threshold ranges. The user is cautioned that this is
subject to both false positive and false negatives; examining diagnostics
plots directly is advised.
Calls \code{\link{diagnostic_warnings.adaptspecfit}}.
}}

\seealso{
\code{\link{time_varying_mean_mean.adaptspecmixturefit}} for estimates of the
time-varying mean and spectral density.
}
