% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptspecx.R
\name{adaptspecx}
\alias{adaptspecx}
\title{AdaptSPEC-X: Covariate Dependent Spectral Modeling of Multiple Nonstationary Time Series}
\usage{
adaptspecx(
  n_loop,
  n_warm_up,
  data,
  design_matrix,
  n_components,
  detrend = FALSE,
  spline_covariates = rep(TRUE, ncol(design_matrix)),
  component_model = adaptspec_model(),
  mixture_prior = list(mean = 0, precision = 1/100, tau_prior_a_squared = 100,
    tau_prior_nu = 3, tau_prior_upper = 10000, n_spline_bases = 0, spline_type =
    "thinplate"),
  component_tuning = adaptspec_tuning(),
  lsbp_tuning = list(n_swap_moves = 1, swap_move_length = 3),
  start = list(beta = NULL, tau_squared = NULL, categories = NULL, components = NULL,
    x_missing = NULL),
  thin = list(beta_lsbp = 1, tau_squared_lsbp = 1, categories = 1, n_segments = 1, beta
    = 1, tau_squared = 1, cut_points = 1, mu = 1, log_posterior = 1, x_missing = 1),
  show_progress = FALSE,
  run_diagnostics = TRUE
)
}
\arguments{
\item{n_loop}{Number of MCMC iterations to perform}

\item{n_warm_up}{Number of warm-up iterations to discard, so that the
number of samples returned is n_loop - n_warm_up}

\item{data}{Numeric vector or matrix of data. If a matrix, the model assumes
it has multiple independent observations of the same random process, one per
column.}

\item{design_matrix}{Design matrix of covariates. There should be one row per
column of \code{data}.}

\item{n_components}{The number of mixture components at which to truncate the
infinite mixture.}

\item{detrend}{Whether to remove a mean and linear trend from the time-series
prior to fitting. If \code{data} is a matrix, this will be done independently for
each column.}

\item{spline_covariates}{A logical vector of length \code{ncol(design_matrix)}
indicating which covariates to include in the spline (if any).}

\item{component_model}{The AdaptSPEC model to be used for each mixture
component. See \code{\link{adaptspec_model}}.}

\item{mixture_prior}{The prior on the mixture component parameters.}

\item{component_tuning}{Tuning parameters for the AdaptSPEC component MCMC
steps; see \code{\link{adaptspec_tuning}}}

\item{lsbp_tuning}{Tuning parameters for the LSBP sampler. You can set the
number of swap moves (\code{n_swap_moves}) to attempt on each iteration, and
how far to consider swapping components (\code{swap_move_length}).}

\item{start}{List of starting values; either an object returned from this
function, or a list. Values that are NULL/missing are ignored.}

\item{thin}{A list specifying how to thin each output of the MCMC sampler.
A value of 1 indicates no thinning, 2 keeps every second sample, and so on.}

\item{show_progress}{Whether to show a progress indicator during MCMC}

\item{run_diagnostics}{Whether to run diagnostics afterwards to help
determine whether the chain has reached convergence.}
}
\value{
An object of class \code{adaptspeclsbpmixturefit}. The MCMC samples
are available under the following list entries, most of which are of class
\code{\link[coda]{mcmc}} or \code{\link[acoda]{mcmca}}:
\itemize{
  \item \code{beta}: Three dimensional array with samples of coefficients
  for the covariate dependent logistic stick breaking prior, including
  spline coefficients if used.
  \item \code{tau_squared}: Matrix of samples of scale parameter for splin
  coefficients, if used.
  \item \code{components}: List of objects of class
  \code{\link[=adaptspec]{adaptspecfit}} containing the samples for
  each mixture component
  \item \code{x_missing}: List of numeric matrices, where each matrix holds
  the samples of missing values for each column in `data`
  \item \code{log_posterior}: Numeric vector of unnormalised values of the
  log posterior in each iteration.
}
}
\description{
This function implements methodology for analyzing multiple possibly
non-stationary time series that are indexed by covariates using a covariate
dependent infinite mixture. The mixture components are based on AdaptSPEC
It is described by \href{https://arxiv.org/abs/1908.06622}{Bertolacci et al., 2020}.
}
\examples{
# Running adaptspecx with i.i.d. normal data
y <- matrix(rnorm(10000), ncol = 10)
design_matrix <- matrix(1, nrow = 10)
fit <- adaptspecx(
  5000,
  1000,
  y,
  design_matrix,
  n_components = 5
)
}
\seealso{
\code{\link{segment_log_spectra_mean}} for methods calculating
estimates of the spectral densities in each segments.
\code{\link{time_varying_spectra_mean.adaptspecmixturefit}} for estimates of
the time-varying spectral density. \code{\link{adaptspecmixturefit}} for
other useful methods applying to objects returned by this function.
\code{\link{merge_samples}} to merge samples from independent runs of this
function.
\code{\link{component_probabilities}} to calculate the probabilities of
component membership.
}
