% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diagnostics}
\alias{diagnostics}
\alias{diagnostic_plots}
\alias{diagnostic_warnings}
\title{Generics for diagnostic methods}
\usage{
diagnostics(x, ...)

diagnostic_plots(x, ...)

diagnostic_warnings(x, ...)
}
\arguments{
\item{x}{The object to perform the diagnostics for.}

\item{...}{Extra argument, possibly used by class-specific mixtures.}
}
\description{
These generics perform various diagnostic routines for AdaptSPEC and
AdaptSPEC-X outputs. See \code{\link{adaptspecfit}}
or \code{\link{adaptspecmixturefit}} for the class-specific methods.
}
\section{Functions}{
\itemize{
\item \code{diagnostic_plots}: Create diagnostic plots.

\item \code{diagnostic_warnings}: Give diagnostic warnings.
}}

