% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intracranial_eeg_data.R
\docType{data}
\name{intracranial_eeg}
\alias{intracranial_eeg}
\title{Intracranial Electroencephalograph (IEEG) Dataset}
\format{A vector time series of 6,000 observations of intracranial electroencephalograph}
\source{
kaggle.com
}
\usage{
data(intracranial_eeg)
}
\description{
A sample of IEEG data from a subject in an interictal state.
}
\references{
https://www.kaggle.com/c/melbourne-university-seizure-prediction
}
\keyword{datasets}
