% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge samples from AdaptSPEC or AdaptSPEC-X}
\usage{
merge_samples(values)
}
\arguments{
\item{values}{List of objects to merge, either from \code{\link{adaptspec}}
or \code{\link{adaptspecx}}}
}
\value{
The merged object
}
\description{
This function merges samples from different runs of AdaptSPEC or AdaptSPEC-X
into a single object. This is used to pool samples from MCMC runs made in
parallel.
}
