% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ntimes}
\alias{ntimes}
\title{Generic for the number of time periods in the input}
\usage{
ntimes(x, ...)
}
\arguments{
\item{x}{The object to return the number of times for}

\item{...}{Extra argument, possibly used by class-specific mixtures.}
}
\description{
Generic function to return the numbers of time periods in the input for
either an AdaptSPEC or AdaptSPEC-X object. See \code{\link{adaptspecfit}}
or \code{\link{adaptspecmixturefit}} for the class-specific methods.
}
