% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-adaptspecfit.R
\name{plot.adaptspecfit}
\alias{plot.adaptspecfit}
\title{Summary plots of adaptspecfit objects}
\usage{
plot.mcmc(fit, ask = dev.interactive(), auto_layout = TRUE)
}
\arguments{
\item{fit}{The fit object, as returned by adaptspec}

\item{ask}{Prompt user before each page of plots}

\item{auto_layout}{Automatically generate output format}
}
\description{
Plots an adaptspecfit object
}

