% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptspecfit.R
\name{segment_log_spectra_mean}
\alias{segment_log_spectra_mean}
\alias{segment_spectra_mean}
\title{Estimates of segment spectral densities from adaptspecfit objects}
\usage{
segment_log_spectra_mean(fit, n_frequencies = 64, frequencies = seq(0, 0.5,
  length.out = n_frequencies))

segment_spectra_mean(fit, n_frequencies = 64, frequencies = seq(0, 0.5,
  length.out = n_frequencies))
}
\arguments{
\item{fit}{\code{adaptspecfit} object}

\item{n_frequencies}{Number of frequencies at which to evaluate the spectral
densities}

\item{frequencies}{Frequencies at which to evaluate the spectral density.
Should be between 0 and 0.5.}
}
\value{
For \code{segment_log_spectra_mean}, a list with entries
\code{frequencies} and \code{log_spectrum}. For \code{segment_spectra_mean},
the entries are \code{frequencies} and \code{spectrum}.
}
\description{
These methods calculates estimates of the segment spectral densities
modelled by \code{\link{adaptspec}}, as evaluated at specified frequencies.
\code{segment_log_spectra_mean} calculates the posterior mean of the log
spectral density, while \code{segment_spectra_mean} does the same for the
spectral density.
}
\section{Functions}{
\itemize{
\item \code{segment_spectra_mean}: Posterior mean of spectral density in
each segment.
}}

