% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{thinplate_spline_basis}
\alias{thinplate_spline_basis}
\title{Thin plate splines}
\usage{
thinplate_spline_basis(covariates, n_bases = floor(nrow(covariates)/10),
  order = 2, omit_intercept = FALSE)
}
\description{
Thin plate splines
}
