% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_varying_mean_mean}
\alias{time_varying_mean_mean}
\alias{time_varying_mean_samples}
\alias{time_varying_spectra_mean}
\alias{time_varying_spectra_samples}
\title{Generics for calculating the time varying mean and spectrum}
\usage{
time_varying_mean_mean(x, ...)

time_varying_mean_samples(x, ...)

time_varying_spectra_mean(x, ...)

time_varying_spectra_samples(x, ...)
}
\arguments{
\item{x}{The object to calculate the quantity for.}

\item{...}{Extra argument, possibly used by class-specific mixtures.}
}
\description{
These generics perform calculate quantities related to the time varying mean
and spectra for fits made with AdaptSPEC and AdaptSPEC-X outputs. See
See \code{\link{time_varying_mean_mean.adaptspecfit}} or
\code{\link{time_varying_mean_mean.adaptspecmixturefit}} for class-specific
methods.
}
\section{Functions}{
\itemize{
\item \code{time_varying_mean_samples}: Posterior samples from the time varying
mean.

\item \code{time_varying_spectra_mean}: Posterior mean of the time varying
spectra.

\item \code{time_varying_spectra_samples}: Posterior samples from the time varying
spectra.
}}

