% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptspecfit.R
\name{time_varying_mean_samples.adaptspecfit}
\alias{time_varying_mean_samples.adaptspecfit}
\title{Samples of the time varying mean from an adaptspecfit object}
\usage{
\method{time_varying_mean_samples}{adaptspecfit}(fit, time_step = 1,
  times = seq(1, ntimes(fit), by = time_step))
}
\arguments{
\item{fit}{\code{adaptspecfit} object}

\item{time_step}{Time varying mean is calculated only at times
divisible by this number. Reduces the size of the output. Ignore if
\code{times} is provided.}

\item{times}{Times at which to calculate the mean. Must be between
1 and \code{ntimes(fit)}, inclusive.}
}
\value{
A matrix. First dimension is sample, second dimension is time.
Attribute 'times' contain the corresponding times (use \code{str()} to
inspect this object to see).
}
\description{
This method calculates samples from the time varying mean modelled by
\code{\link{adaptspec}}. If \code{segment_means = FALSE} for the fit,
the mean will be zero everywhere.
}
