% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptspecfit.R
\name{adaptspecfit}
\alias{adaptspecfit}
\alias{window.adaptspecfit}
\alias{summary.adaptspecfit}
\alias{ntimes.adaptspecfit}
\alias{diagnostic_plots.adaptspecfit}
\alias{diagnostics.adaptspecfit}
\alias{diagnostic_warnings.adaptspecfit}
\alias{plot.adaptspecfit}
\title{Methods for adaptspecfit objects}
\usage{
\method{window}{adaptspecfit}(fit, ...)

\method{summary}{adaptspecfit}(fit, iterations_threshold = 0)

\method{ntimes}{adaptspecfit}(fit)

\method{diagnostic_plots}{adaptspecfit}(fit, ...)

\method{diagnostics}{adaptspecfit}(fit, iterations_threshold = 0)

\method{diagnostic_warnings}{adaptspecfit}(fit,
  effective_size_threshold = 100, iterations_proportion_threshold = 0.01,
  prefix = "")

\method{plot}{adaptspecfit}(fit, ask, auto_layout = TRUE)
}
\arguments{
\item{fit}{\code{adaptspecfit} object}

\item{iterations_threshold}{Number of iterations below which to suppress
output. For example, if this is 100, and the MCMC chain spent less than
100 iterations in, say, n_segments = 4, summary output will not include
n_segments = 4}

\item{effective_size_threshold}{Minimum effective sample size (see
\code{\link[coda]{effectiveSize}}) below which a warning is output.}

\item{iterations_proportion_threshold}{Proportion of iterations below which
to suppress output. Just like \code{iterations_threshold}, but a proportion.}

\item{prefix}{Used internally.}

\item{ask}{Prompt user before each page of plots}

\item{auto_layout}{Automatically generate output format}
}
\description{
These methods apply to the adaptspecfit objects returned by
\code{\link{adaptspec}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{window}: Method to modify the start/thinning of MCMC samples,
as per \code{\link[coda]{window.mcmc}}

\item \code{summary}: Summarises MCMC samples.

\item \code{ntimes}: Returns the number of time periods in the input

\item \code{diagnostic_plots}: MCMC diagnostic plots to assess convergence.

\item \code{diagnostics}: Outputs MCMC diagnostics statistics to help assess
convergence.

\item \code{diagnostic_warnings}: Outputs warnings when MCMC diagnostics are below
outside of nominal threshold ranges. The user is cautioned that this is
subject to both false positive and false negatives; examining diagnostics
plots directly is advised.

\item \code{plot}: Summary plots
}}

