% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptspecfit.R
\name{time_varying_spectra_mean.adaptspecfit}
\alias{time_varying_spectra_mean.adaptspecfit}
\title{Posterior mean estimate of the time varying spectral density from
an adaptspecfit object}
\usage{
\method{time_varying_spectra_mean}{adaptspecfit}(fit, n_frequencies = 64,
  time_step = 1, frequencies = seq(0, 0.5, length.out = n_frequencies),
  times = seq(1, ntimes(fit), by = time_step))
}
\arguments{
\item{fit}{\code{adaptspecfit} object}

\item{n_frequencies}{Number of frequencies at which to evaluate the spectral
densities. Ignored if \code{frequencies} is set.}

\item{time_step}{Time varying spectral density is calculated only at times
divisible by this number. Reduces the size of the output. Ignored if
\code{times} is set.}

\item{frequencies}{Frequencies at which to evaluate the spectral density.
Must be between 0 and 0.5, inclusive.}

\item{times}{Times at which to calculate the spectra density. Must be between
1 and \code{ntimes(fit)}, inclusive.}
}
\value{
Numeric matrix. Rows hold frequencies, columns hold times. Attributes
'frequencies' and 'times' contain the corresponding frequencies/times
(use \code{str()} to inspect this object to see).
}
\description{
This method calculates the posterior mean of the time varying spectral
density modelled by \code{\link{adaptspec}}. This function can take a lot of
time and memory, so consider thinning the input \code{fit} using
\code{\link{window.adaptspecfit}} prior to calling, or using the
\code{time_step} argument.
}
\details{
A matrix with frequency on the rows and times on the
columns. Attributes 'frequencies' and 'times' contain the corresponding
frequencies and times.
}
\examples{
data(simulated_piecewise)
fit <- adaptspec(
  5000,
  1000,
  simulated_piecewise,
  n_bases = 15
)
tvsm <- time_varying_spectra_mean(window(fit, thin = 10), 128, 5)
image(
  t(tvsm),
  x = attr(tvsm, 'times'),
  y = attr(tvsm, 'frequencies'),
  xlab = 'Time',
  ylab = 'Frequency',
  col = terrain.colors(50)
)
}
\seealso{
\code{time_varying_spectra_samples} for samples of the time-varying
spectrum
}
