% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptspecfit.R
\name{time_varying_spectra_samples.adaptspecfit}
\alias{time_varying_spectra_samples.adaptspecfit}
\title{Samples of the time varying spectral density from an adaptspecfit object}
\usage{
\method{time_varying_spectra_samples}{adaptspecfit}(fit, n_frequencies = 64,
  time_step = 1, frequencies = seq(0, 0.5, length.out = n_frequencies),
  times = seq(1, ntimes(fit), by = time_step))
}
\arguments{
\item{fit}{\code{adaptspecfit} object}

\item{n_frequencies}{Number of frequencies at which to evaluate the spectral
densities. Ignored if \code{frequencies} is set.}

\item{time_step}{Time varying spectral density is calculated only at times
divisible by this number. Reduces the size of the output. Ignored if
\code{times} is set.}

\item{frequencies}{Frequencies at which to evaluate the spectral density.
Must be between 0 and 0.5, inclusive.}

\item{times}{Times at which to calculate the spectra density. Must be between
1 and \code{ntimes(fit)}, inclusive.}
}
\value{
Three dimensional array. First dimension is sample, second
dimension is frequency, and third dimension is time. Attributes 'frequencies'
and 'times' contain the corresponding frequencies/times (use \code{str()}
to inspect this object to see).
}
\description{
This method calculates samples from the time varying spectral density
modelled by \code{\link{adaptspec}}. This function can take a lot of time
and memory, so consider thinning the input \code{fit} using
\code{\link{window.adaptspecfit}} prior to calling, or using the
\code{time_step} argument.
}
\details{
A matrix with frequency on the rows and times on the
columns. Attributes 'frequencies' and 'times' contain the corresponding
frequencies and times.
}
